Deprecate the RUI install script with a helpful replacement message
It turns out that you cannot simply copy virtualenv folders around and then delete the original source -- they point back to that source. This makes it necessary to set up the RUI virtualenv from the folder from which it is ultimately intended to run (typically /opt/numurus/ros/share/numurus_rui). The install script did not do this.

While it isn't strictly necessary to build the entire RUI front-end from the installation directory, it is good practice. Ideally the RUI front-end build should be scripted and that can be run manually or automatically as a Catkin post-install step as necessary.