Handle nodes without Display Name properly in configuration changes
Previously, if a node did not have a defined "Display Name" (via displayname json file), any attempt to save or reset its config via RUI would fail. However, in that case, the node as it is known in the Subsystem drop-down is already the correct ROS node name, so we fall back to just using that.