Add (optionally hidden) image quality selector to CameraViewer panels
This is a major part of a bandwidth reduction strategy for the RUI. The checkboxes allow the user to set a different "quality" parameter for the web_video_server-based image sources. Selecting "Low" reduces bandwidth consumption considerably by setting quality=10 (jpeg compression value). It also reduces frame rate and increases latency due to computational burden of compressing and decompressing the image.