Bug Fix: Properly convert detection_threshold to a number
When loading the classifier page, if a classifier is already running, the threshold value is taken from what that classifier reported. However, it was being improperly converted to a string (via a rounding function). If a user later tried to stop and then restart the classifer without changing the slider bar value, the string was transmitted and the message was rejected by rosbridge.

This commit simply converts the threshold back to a number after the rounding function by way of the + operator (as we do other places in code).