Bug fix: Eliminate race case exposed by removing login requirement
When the login requirement was removed, it exposed a race case that the array of 3DX sensors is not populated (via ROS topic introspection) by the time the 3DX page loads when navigating directly to that page. This commit primarily fixes that.

Note that now the user must actively select the 3DX sensor from the drop-down when navigating directly to 3DX page

This commit also removes a bunch of extra login cruft. That was done during debugging (ended up not helping the issue), but is a reasonable code clean-up step, so I'm committing it here.