3DX Tab: Only auto-select sensor if there is exactly one option
RUI-31

This is largely a work-around for an issue with the CameraViewer not resetting the canvas size when a new sensor is selected. So when sonar is auto-selected, then camera is selected the camera viewer has invalid size.

Users will still have an issue if they try to switch between sensors via drop-down. Will need to refresh page instead.