Update image aspect ratio on new sensor selection
RUI-31
RUI-3 (partial fix)

Now the canvas size will be updated when a new sensor (or new image from that sensor) is selected.

This does not address the other issue of the camera images looking weird when crop angle (3DX angle) is adjusted... that would require the Sensor3DX parent of CameraViewer to somehow inform CameraViewer of the angle change and then CameraViewer adjusting the parameters to context.drawImage() to crop the image in the canvas. But that is a bit awkward as it is only relevant for angle changes on camera (angle changes on sonar would be cropped differently if at all, since they impact TX angle).