Bug fix: Avoid creating and attaching new Image objects when navigating away from Image viewer page
Merge Candidate. RUI-15. RUI-28.

Previously, the assignment of image.src to the web_video_server URL was happening through a callstack that originated at onCanvasRef, which was being called both when mounting and unmounting the CameraViewer. But this leads to a bunch of dangling Image objects that aren't properly cleaned up. The fix was to only do the image.src assigment when the component mounts (via React's componentDidMount()).

This also appears to fix the issue where image topics are not properly unsubscribed when no longer needed.