Fix a bug with button style updates not triggering rerender
This was introduced when Button was promoted to full React Component (class instead of function) recently for the visual-feedback-on-push update.

Because ColoredIndicator and BooleanIndicator are Button objects under the hood whose background style changes to provide visualization, these widgets were broken.

Root cause here is that prop updates don't trigger render() on components, only state updates do, so we use a state tied to a prop and the componentDidUpdate() method to fix the issue.