Bug fix: Properly show imagery if classifier already running when page loads
This commit still leaves (at least) one dangling issue: If the Detection page is reloaded the props.ros does not yet exist at constructor time, so image is not properly updated.

This can probably be solved via judicious use of componentDidUpdate() but really this entire page's use of state and props seems like an anti-pattern and the entire logic could stand to be restructured... it has outgrown its initial use case.