Add a progress bar for A/I model loading
NE-13

Progress bar only appears during the “Loading” phase to avoid screen clutter. The rate that the progress bar increments is a (affine) function only of the size of the weights file for the classifier model, and I only currently have two models to test, so an updated linear slope (or some other non-linear curve) may have to be implemented as we look at other models of other sizes. The true loading time is probably also a function of what else is going on on the system CPU and GPU-wise at a give time, but that is not part of our progress calculation here.