Detection: Tie displayed image directly to classifier state
NE-137

Biggest change here is that we are now always monitoring the classifier state and showing the classifier image whenever the classifier is running. When not running, the image will display whatever the user selection is (or was prior to a classifier session).

This allows identical display behavior between starting/stopping classifier via RUI or externally (via a different RUI or ROS connection).