Improve Connect page responsiveness
Almost everything displayed on Connect page is updated via the nepi_link_status_query service, which runs periodically.

This commit introduces a one-shot synchronous options for that service call so that when a change is applied by the user, the service is called immediately.

This also allows us to do the periodic call much less frequently, so less overhead.